major refactor: Removed dietPlan info from DailyMeals to improve cohesion/reduce coupling Removed DietPlanWrapper class to simplify code Created DietController class to reduce coupling with front and backend, and to allow recommendation functionality in future updates Created more packages to seperate classes
