+++ app/src/main/java/com/ganterpore/simplediet/Controller/DailyMeals.java
@@ -4,7 +4,6 @@ import android.support.annotation.NonNull;
 import android.text.format.DateUtils;
 import android.util.Log;
 
-import com.ganterpore.simplediet.Model.DietPlan;
 import com.ganterpore.simplediet.Model.Meal;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -24,11 +23,17 @@ import java.util.List;
 
 import javax.annotation.Nullable;
 
-public class DailyMeals implements DietPlanWrapper.DietPlanInterface {
+public class DailyMeals {
     private static final String TAG = "DailyMeals";
+    public static final String MEALS = "Meals";
     private FirebaseFirestore db;
     private List<DailyMealsInterface> listeners;
-    private List<Meal> meals;
+
+    private List<Meal> todaysMeals;
+//    private List<Meal> yesterdaysMeals;
+//    private List<Meal> theDayBeforesMeals;
+    private List<Meal> thisWeeksMeals;
+//    private List<Meal> tomorrowsMeals;
 
     private double vegCount;
     private double proteinCount;
@@ -38,12 +43,10 @@ public class DailyMeals implements DietPlanWrapper.DietPlanInterface {
     private double waterCount;
     private double excessServes;
 
-    private double totalCheats;
-    private double weeklyCheats;
+    private double totalCheats; //accumulated cheats for this day
+    private double weeklyCheats; //accumulated cheats for the past week
 
     private Date date;
-
-    private DietPlanWrapper dietPlan;
 
     /**
      * Constructor for getting todays meals
@@ -69,101 +72,110 @@ public class DailyMeals implements DietPlanWrapper.DietPlanInterface {
         listeners = new ArrayList<>();
         addListener(listener);
 
-        //get tomorrows date
-        Date nextDay = new Date(day.getTime() + DateUtils.DAY_IN_MILLIS);
-        Date weekAgo = new Date(day.getTime() - DateUtils.DAY_IN_MILLIS*7);
-        //update the days to the start of each day
+        //update date to start of the day, and get other important dates
         day = getStartOfDay(day);
-        nextDay = getStartOfDay(nextDay);
-        weekAgo = getStartOfDay(weekAgo);
+        final Date nextDay = new Date(day.getTime() + DateUtils.DAY_IN_MILLIS);
+        final Date endOfNextDay = new Date(day.getTime() + 2*DateUtils.DAY_IN_MILLIS);
+        final Date yesterDay = new Date(day.getTime() - DateUtils.DAY_IN_MILLIS);
+        final Date dayBeforeYest = new Date(day.getTime() - 2*DateUtils.DAY_IN_MILLIS);
+        final Date weekAgo = new Date(day.getTime() - 7*DateUtils.DAY_IN_MILLIS);
 
         this.date = day;
 
-        //access documents from the database from between these two times
         db = FirebaseFirestore.getInstance();
-        CollectionReference meals = db.collection("Meals");
-        Query todaysMeals = meals.whereGreaterThanOrEqualTo("day", day.getTime())
-                .whereLessThan("day", nextDay.getTime())
-                .whereEqualTo("user", FirebaseAuth.getInstance().getCurrentUser().getUid());
-
-        //access documents for meals from the past week
-        Query lastWeeksMeals = meals.whereGreaterThanOrEqualTo("day", weekAgo.getTime())
-                .whereLessThan("day", nextDay.getTime())
-                .whereEqualTo("user", FirebaseAuth.getInstance().getCurrentUser().getUid());
+        CollectionReference mealsCollection = db.collection(MEALS);
+        Query allMeals = mealsCollection.whereGreaterThan("day", weekAgo.getTime())
+                            .whereLessThan("day", endOfNextDay.getTime())
+                            .whereEqualTo("user", FirebaseAuth.getInstance().getCurrentUser().getUid());
 
         //add a snapshot listener to update the meals when the databse updates
-        todaysMeals.addSnapshotListener(new EventListener<QuerySnapshot>() {
+        final Date finalDay = day;
+        allMeals.addSnapshotListener(new EventListener<QuerySnapshot>() {
             @Override
             public void onEvent(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable FirebaseFirestoreException e) {
                 if(e != null) {
                     Log.e(TAG, "onEvent: exception " + e.getLocalizedMessage());
                 } else {
                     List<DocumentSnapshot> meals = queryDocumentSnapshots.getDocuments();
-                    getDataFromDocuments(meals);
+                    sortMeals(meals, finalDay, nextDay, yesterDay, dayBeforeYest, weekAgo, endOfNextDay);
+                    updateData();
                     updateListeners();
                 }
             }
         });
+
         //get the query, and update when data recieved
-        todaysMeals.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+        allMeals.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
             @Override
             public void onComplete(@NonNull Task<QuerySnapshot> task) {
                 if(task.isSuccessful()) {
                     List<DocumentSnapshot> meals = task.getResult().getDocuments();
-                    getDataFromDocuments(meals);
+                    sortMeals(meals, finalDay, nextDay, yesterDay, dayBeforeYest, weekAgo, endOfNextDay);
+                    updateData();
                     updateListeners();
                 }
             }
         });
+//        dietPlan = new DietPlanWrapper(this, user);
+    }
 
-        //updating the number of cheat points consumed in the past week
-        lastWeeksMeals.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                if(task.isSuccessful()) {
-                    List<DocumentSnapshot> meals = task.getResult().getDocuments();
-                    weeklyCheats = 0;
-                    for(DocumentSnapshot meal : meals) {
-                        double serveCount = 0;
-                        double cheatScore = meal.getDouble("cheatScore");
-                        double mealVegCount  = meal.getDouble("vegCount");
-                        double mealProteinCount = meal.getDouble("proteinCount");
-                        double mealDairyCount = meal.getDouble("dairyCount");
-                        double mealGrainCount = meal.getDouble("grainCount");
-                        double mealFruitCount = meal.getDouble("fruitCount");
-                        double mealWaterCount = meal.getDouble("waterCount");
-                        double mealExcessServes = meal.getDouble("excessServes");
-
-                        serveCount = mealVegCount + mealProteinCount + mealDairyCount
-                                + mealGrainCount + mealFruitCount + mealWaterCount + mealExcessServes;
-                        weeklyCheats += cheatScore * serveCount;
-                    }
-
-                    updateListeners();
-                }
+    private void sortMeals(List<DocumentSnapshot> meals, Date finalDay, Date nextDay, Date yesterDay, Date dayBeforeYest, Date weekAgo, Date endOfNextDay) {
+        todaysMeals = new ArrayList<>();
+//        yesterdaysMeals =  new ArrayList<>();
+//        theDayBeforesMeals = new ArrayList<>();
+        thisWeeksMeals = new ArrayList<>();
+//        tomorrowsMeals = new ArrayList<>();
+        for(DocumentSnapshot mealDS : meals) {
+            Meal meal = new Meal(mealDS);
+            if(meal.getDay() >= finalDay.getTime() && meal.getDay() < nextDay.getTime()) {
+                todaysMeals.add(meal);
             }
-        });
-
-        dietPlan = new DietPlanWrapper(this, user);
+//            if(meal.getDay() >= yesterDay.getTime() && meal.getDay() < finalDay.getTime()) {
+//                yesterdaysMeals.add(meal);
+//            }
+//            if(meal.getDay() >= dayBeforeYest.getTime() && meal.getDay() < yesterDay.getTime()) {
+//                theDayBeforesMeals.add(meal);
+//            }
+            if(meal.getDay() >= weekAgo.getTime() && meal.getDay() < nextDay.getTime()) {
+                thisWeeksMeals.add(meal);
+            }
+//            if(meal.getDay() >= nextDay.getTime() && meal.getDay() < endOfNextDay.getTime()) {
+//                tomorrowsMeals.add(meal);
+//            }
+        }
     }
 
-    //checks if enough water has been consumed for this day
-    public boolean isWaterCompleted() {
-        return waterCount >= dietPlan.getDietPlan().getDailyWater();
-    }
-
-    //checks if enough food has been consumed for this dat
-    public boolean isFoodCompleted() {
-        return vegCount >= dietPlan.getDietPlan().getDailyVeges()
-                & proteinCount >= dietPlan.getDietPlan().getDailyProtein()
-                & dairyCount >= dietPlan.getDietPlan().getDailyDairy()
-                & grainCount >= dietPlan.getDietPlan().getDailyGrain()
-                & fruitCount >= dietPlan.getDietPlan().getDailyFruit();
-    }
-
-    //checks if have gone over the weekly cheat amount allowable
-    public boolean isOverCheatScore() {
-        return weeklyCheats > dietPlan.getDietPlan().getWeeklyCheats();
+    /**
+     * given a list of documents of all the docMeals in the day, update the data
+     * to reflect what is in the documents
+     */
+    private void updateData() {
+        //reset all numbers to 0
+        vegCount = 0;
+        proteinCount = 0;
+        dairyCount = 0;
+        grainCount = 0;
+        fruitCount = 0;
+        waterCount = 0;
+        excessServes = 0;
+        totalCheats = 0;
+        //iterate through all todaysMeals and add the days data
+        for(Meal meal : todaysMeals) {
+            totalCheats += meal.calculateTotalCheats();
+            vegCount += meal.getVegCount();
+            proteinCount += meal.getProteinCount();
+            dairyCount += meal.getDairyCount();
+            grainCount += meal.getGrainCount();
+            fruitCount += meal.getFruitCount();
+            waterCount += meal.getWaterCount();
+            excessServes += meal.getExcessServes();
+        }
+        weeklyCheats = 0;
+        int i =0;
+        for(Meal meal : thisWeeksMeals) {
+            i++;
+            weeklyCheats += meal.calculateTotalCheats();
+        }
     }
 
     public void addListener(DailyMealsInterface listener) {
@@ -172,8 +184,23 @@ public class DailyMeals implements DietPlanWrapper.DietPlanInterface {
         }
     }
 
+    /**
+     * get a date representing the start time of a given day
+     * @param date, the date with which you want the start time of the day from
+     * @return the date at the start of the given day
+     */
+    private Date getStartOfDay(Date date) {
+        Calendar calendar = Calendar.getInstance();
+        calendar.setTime(date);
+        int year = calendar.get(Calendar.YEAR);
+        int month = calendar.get(Calendar.MONTH);
+        int day = calendar.get(Calendar.DATE);
+        calendar.set(year, month, day, 0, 0, 0);
+        return calendar.getTime();
+    }
+
     public List<Meal> getMeals() {
-        return meals;
+        return todaysMeals;
     }
 
     public double getTotalServes() {
@@ -217,52 +244,8 @@ public class DailyMeals implements DietPlanWrapper.DietPlanInterface {
         return date;
     }
 
-    public void setDate(Date date) {
-        this.date = date;
-    }
-
-    /**
-     * given a list of documents of all the docMeals in the day, update the data
-     * to reflect what is in the documents
-     * @param docMeals, a list of FirestoreDcument's of the days docMeals
-     */
-    private void getDataFromDocuments(List<DocumentSnapshot> docMeals) {
-        meals = new ArrayList<>();
-        for(DocumentSnapshot docMeal : docMeals) {
-            meals.add(new Meal(docMeal));
-        }
-        //reset all numbers to 0
-        vegCount = 0;
-        proteinCount = 0;
-        dairyCount = 0;
-        grainCount = 0;
-        fruitCount = 0;
-        waterCount = 0;
-        excessServes = 0;
-        totalCheats = 0;
-        //iterate through all docMeals and add the days data
-        for(Meal meal : meals) {
-            double serveCount = 0;
-            double cheatScore = meal.getCheatScore();
-            double mealVegCount  = meal.getVegCount();
-            double mealProteinCount = meal.getProteinCount();
-            double mealDairyCount = meal.getDairyCount();
-            double mealGrainCount = meal.getGrainCount();
-            double mealFruitCount = meal.getFruitCount();
-            double mealWaterCount = meal.getWaterCount();
-            double mealExcessServes = meal.getExcessServes();
-
-            serveCount = mealVegCount + mealProteinCount + mealDairyCount
-                    + mealGrainCount + mealFruitCount + mealWaterCount + mealExcessServes;
-            totalCheats += cheatScore * serveCount;
-            vegCount += mealVegCount;
-            proteinCount += mealProteinCount;
-            dairyCount += mealDairyCount;
-            grainCount += mealGrainCount;
-            fruitCount += mealFruitCount;
-            waterCount += mealWaterCount;
-            excessServes += mealExcessServes;
-        }
+    public double getWeeklyCheats() {
+        return weeklyCheats;
     }
 
     /**
@@ -274,11 +257,6 @@ public class DailyMeals implements DietPlanWrapper.DietPlanInterface {
         }
     }
 
-    @Override
-    public void updateDietPlan(DietPlan diet) {
-        updateListeners();
-    }
-
     public interface DailyMealsInterface {
         /**
          * This method is called whenever the day object is updated
@@ -288,18 +266,5 @@ public class DailyMeals implements DietPlanWrapper.DietPlanInterface {
     }
 
 
-    /**
-     * get a date representing the start time of a given day
-     * @param date, the date with which you want the start time of the day from
-     * @return the date at the start of the given day
-     */
-    private Date getStartOfDay(Date date) {
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTime(date);
-        int year = calendar.get(Calendar.YEAR);
-        int month = calendar.get(Calendar.MONTH);
-        int day = calendar.get(Calendar.DATE);
-        calendar.set(year, month, day, 0, 0, 0);
-        return calendar.getTime();
-    }
+
 }
